+++ jest.config.js
@@ -11,7 +11,7 @@ module.exports = {
   coverageReporters: ["lcov", "text"],
   coverageThreshold: {
     global: {
-      branches: 75,
+      branches: 80,
       functions: 90,
       lines: 90,
       statements: 90,