+++ test/main_ws.test.js
@@ -112,6 +112,11 @@ describe("Main", () => {
     expect(result).toBeNull();
   });
 
+  test("Event Exception", async () => {
+    const result = await emitEvent(socket, "eventException", {});
+    expect(result).toBeNull();
+  });
+
   test("Event", async () => {
     const eventResultPromise = waitForEvent(socketOther, "customEvent");
     const result = await emitEvent(socket, "triggerCustomEvent", { num: 1, text: "test" });
@@ -176,6 +181,53 @@ describe("Main", () => {
     await eventNoOtherResultPromise;
 
     await enterContext(socket, ID);
+  });
+
+  test("Event Context Static", async () => {
+    const context = "context";
+
+    await enterContext(socket, context);
+    await wait();
+    let eventResultPromise = waitForEvent(socket, "customContextStaticEvent");
+    let eventNoResultPromise = waitForNoEvent(socketOther, "customContextStaticEvent");
+    let result = await emitEvent(socket, "triggerCustomContextStaticEvent", { num: 1, text: "test" });
+    expect(result).toBeNull();
+    let eventResult = await eventResultPromise;
+    expect(eventResult.text).toBe("test1");
+    await eventNoResultPromise;
+
+    await enterContext(socketOther, context);
+    await wait();
+    eventResultPromise = waitForEvent(socket, "customContextStaticEvent");
+    const eventResultOtherPromise = waitForEvent(socketOther, "customContextStaticEvent");
+    result = await emitEvent(socket, "triggerCustomContextStaticEvent", { num: 1, text: "test" });
+    expect(result).toBeNull();
+    eventResult = await eventResultPromise;
+    expect(eventResult.text).toBe("test1");
+    const eventResultOther = await eventResultOtherPromise;
+    expect(eventResultOther.text).toBe("test1");
+
+    await exitContext(socketOther, context);
+    await wait();
+    eventResultPromise = waitForEvent(socket, "customContextStaticEvent");
+    eventNoResultPromise = waitForNoEvent(socketOther, "customContextStaticEvent");
+    result = await emitEvent(socket, "triggerCustomContextStaticEvent", { num: 1, text: "test" });
+    expect(result).toBeNull();
+    eventResult = await eventResultPromise;
+    expect(eventResult.text).toBe("test1");
+    await eventNoResultPromise;
+
+    await exitContext(socket, context);
+    await exitContext(socketOther, context);
+    await wait();
+    eventNoResultPromise = waitForNoEvent(socket, "customContextStaticEvent");
+    const eventNoOtherResultPromise = waitForNoEvent(socketOther, "customContextStaticEvent");
+    result = await emitEvent(socket, "triggerCustomContextStaticEvent", { num: 1, text: "test" });
+    expect(result).toBeNull();
+    await eventNoResultPromise;
+    await eventNoOtherResultPromise;
+
+    await enterContext(socket, context);
   });
 
   test("Event Context Mass", async () => {