+++ test/redis_custom_adapter.test.js
@@ -4,11 +4,6 @@ const cds = require("@sap/cds");
 
 const { connect, disconnect, emitEvent } = require("./_env/util/socket.io");
 const xsenv = require("@sap/xsenv");
-const redis = require("redis");
-require("@socket.io/redis-streams-adapter");
-
-jest.mock("redis", () => require("./_env/mocks/redis"));
-jest.mock("@socket.io/redis-streams-adapter", () => require("./_env/mocks/redisStreamAdapter"));
 
 jest.spyOn(xsenv, "serviceCredentials").mockReturnValue({ uri: "uri" });
 
@@ -17,7 +12,7 @@ cds.test(__dirname + "/_env");
 cds.env.websocket = {
   kind: "socket.io",
   adapter: {
-    impl: "@socket.io/redis-streams-adapter",
+    impl: __dirname + "/_env/mocks/redisCustomAdapter",
     local: true,
     options: {
       key: "websocket",
@@ -36,12 +31,8 @@ describe("Redis", () => {
     await disconnect(socket);
   });
 
-  test("Redis Adapter", async () => {
+  test("Redis Custom Adapter", async () => {
     const result = await emitEvent(socket, "message", { text: "test" });
     expect(result).toBe("test");
-
-    expect(redis.createClient).toHaveBeenCalledWith({ url: "uri" });
-    expect(redis.client.connect).toHaveBeenCalledWith();
-    expect(redis.client.on).toHaveBeenNthCalledWith(1, "error", expect.any(Function));
   });
 });